feat: Server => Create models for course feature, improve folder distribution
feat: Server => Create models for course feature, improve folder distribution
chore: Server => Each feature has their own folder
