+++ server/controllers/auth.controller.js
@@ -1,4 +1,4 @@
-import { User } from "../models/user.model.js";
+import { User } from "../models/user/user.model.js";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcryptjs";
 import { TOKEN_SECRET, NODE_ENV } from "../config.js";
