+++ server/models/classification/expression.model.js
@@ -1,0 +1,14 @@
+import mongoose from 'mongoose';
+
+export const expressionSchema = new mongoose.Schema(
+    {
+        angry: Number,
+        disgusted: Number,
+        fearful: Number,
+        happy: Number,
+        neutral: Number,
+        sad: Number,
+        surprised: Number,
+    },
+    { _id: false }
+);
